refactor(tasks): move Go2 shared base [skip ci]
