+++ docs/sphinx/source/api_reference/envs/locomotion.md
@@ -9,7 +9,7 @@
    unilab.envs.locomotion.common
    unilab.envs.locomotion.g1
    unilab.envs.locomotion.go1
-   unilab.envs.locomotion.go2
+   unilab.tasks.locomotion.go2
    unilab.envs.locomotion.go2_arm
    unilab.envs.locomotion.go2w
 ```
