+++ src/unilab/envs/locomotion/__init__.py
@@ -2,7 +2,6 @@
 
 __unilab_registry_modules__ = (
     "unilab.envs.locomotion.go1",
-    "unilab.envs.locomotion.go2",
     "unilab.envs.locomotion.go2w",
     "unilab.envs.locomotion.g1",
     "unilab.envs.locomotion.go2_arm",
