--- src/unilab/envs/locomotion/go2/__init__.py
@@ -1,1 +1,0 @@
-"""Legacy Go2 shared base pending migration to :mod:`unilab.tasks`."""
